Overhaul the app version 1.0 for public use
The app is still in its alpha version but, can be used as a time tracker
for your daily freelance task tracking. This is just a proof of concept to
check whether there is a need of such tool since, there are much
better tools out there to experiment with.